+++ rpc/backend/call_tx.go
@@ -127,6 +127,7 @@ func (b *Backend) SendRawTransaction(data hexutil.Bytes) (common.Hash, error) {
 	}
 
 	// Query params to use the EVM denomination
+	// use latest queryClient in send
 	res, err := b.queryClient.QueryClient.Params(b.ctx, &evmtypes.QueryParamsRequest{})
 	if err != nil {
 		b.logger.Error("failed to query evm params", "error", err.Error())
@@ -331,6 +332,7 @@ func (b *Backend) EstimateGas(args evmtypes.TransactionArgs, blockNrOptional *rp
 	// From ContextWithHeight: if the provided height is 0,
 	// it will return an empty context and the gRPC query will use
 	// the latest block height for querying.
+	// use latest queryClient to estimate
 	res, err := b.queryClient.EstimateGas(rpctypes.ContextWithHeight(blockNr.Int64()), &req)
 	if err != nil {
 		return 0, err
@@ -366,7 +368,8 @@ func (b *Backend) DoCall(
 	// From ContextWithHeight: if the provided height is 0,
 	// it will return an empty context and the gRPC query will use
 	// the latest block height for querying.
-	ctx := rpctypes.ContextWithHeight(blockNr.Int64())
+	height := blockNr.Int64()
+	ctx := rpctypes.ContextWithHeight(height)
 	timeout := b.RPCEVMTimeout()
 
 	// Setup context so it may be canceled the call has completed
@@ -381,8 +384,7 @@ func (b *Backend) DoCall(
 	// Make sure the context is canceled when the call has completed
 	// this makes sure resources are cleaned up.
 	defer cancel()
-
-	res, err := b.queryClient.EthCall(ctx, &req)
+	res, err := b.getGrpcClient(height).EthCall(ctx, &req)
 	if err != nil {
 		return nil, err
 	}