+++ rpc/backend/tracing.go
@@ -13,6 +13,17 @@ import (
 	tmrpctypes "github.com/tendermint/tendermint/rpc/core/types"
 )
 
+func (b *Backend) getGrpcClient(height int64) *rpctypes.QueryClient {
+	for blocks, client := range b.backupQueryClients {
+		// b1-b2 -> g1
+		// b3-b4 -> g2
+		if int64(blocks[0]) <= height && int64(blocks[1]) >= height {
+			return client
+		}
+	}
+	return b.queryClient
+}
+
 // TraceTransaction returns the structured logs created during the execution of EVM
 // and returns them as a JSON object.
 func (b *Backend) TraceTransaction(hash common.Hash, config *evmtypes.TraceConfig) (interface{}, error) {
@@ -98,7 +109,8 @@ func (b *Backend) TraceTransaction(hash common.Hash, config *evmtypes.TraceConfi
 		// 0 is a special value in `ContextWithHeight`
 		contextHeight = 1
 	}
-	traceResult, err := b.queryClient.TraceTx(rpctypes.ContextWithHeight(contextHeight), &traceTxRequest)
+	queryClient := b.getGrpcClient(contextHeight)
+	traceResult, err := queryClient.TraceTx(rpctypes.ContextWithHeight(contextHeight), &traceTxRequest)
 	if err != nil {
 		return nil, err
 	}
@@ -174,8 +186,8 @@ func (b *Backend) TraceBlock(height rpctypes.BlockNumber,
 		ProposerAddress: sdk.ConsAddress(block.Block.ProposerAddress),
 		ChainId:         b.chainID.Int64(),
 	}
-
-	res, err := b.queryClient.TraceBlock(ctxWithHeight, traceBlockRequest)
+	queryClient := b.getGrpcClient(int64(contextHeight))
+	res, err := queryClient.TraceBlock(ctxWithHeight, traceBlockRequest)
 	if err != nil {
 		return nil, err
 	}