+++ server/config/toml.go
@@ -83,6 +83,10 @@ return-data-limit = {{ .JSONRPC.ReturnDataLimit }}
 # Upgrade height for fix of revert gas refund logic when transaction reverted.
 fix-revert-gas-refund-height = {{ .JSONRPC.FixRevertGasRefundHeight }}
 
+# A list of backup grpc address with block range
+# Example: "0.0.0.0:26113" = [0, 20]
+backup-grpc-address-block-range = "{{ "{" }}{{ range $k, $v := .JSONRPC.BackupGRPCBlockAddressBlockRange }}\"{{ $v }}\": [{{index $k 0 }}, {{ index $k 1}}]{{ end }}{{ "}" }}"
+
 ###############################################################################
 ###                             TLS Configuration                           ###
 ###############################################################################